# perf: make `ape plugins --help` faster