### What I did

Before:

```
ape plugins --help  0.60s user 0.08s system 99% cpu 0.684 total
```

After:

```
ape plugins --help  0.30s user 0.04s system 98% cpu 0.346 total
```

### How I did it

imports

### How to verify it

<!-- Discuss any methods that should be used to verify the change -->

### Checklist

<!-- All PRs must complete the following checklist before being merged -->

- [ ] All changes are completed
- [ ] New test cases have been added
- [ ] Documentation has been updated